Implement super smart logging
The idea is to have one master logger that is the only one who can
actually writes to the system logs.

All of our logged in players will have a regular logger in their Client
type. Each logger has a max size. When the stash of the logger is full
or when the player logs out the logger writes back his records to the
master logger.

When the master's stash is full, writes back to the system logs. All of
the logging is happening in a seperate goroutine (which is not something
the user of it should think about), but the exact timestamp is preserved
and all the records are sorted by it right before dumping to the system
log.

For now I will just put that there and will write benchmark tests for
it. Because the idea of it is fun, but we can't be sure this won't
appear to be actually slower :) If it's faster I will add levels of
logging (request, response, warning, error, ...) and default prefixes
(e.g. username of the player who logged that).